requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.1.0
django-simple-history==3.3.0
djangorestframework==3.14.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
